+++ src/elements/ContactElement.php
@@ -455,7 +455,7 @@ class ContactElement extends Element
      */
     public function getPostEditUrl(): ?string
     {
-        return UrlHelper::cpUrl("campaign/contacts");
+        return UrlHelper::cpUrl('campaign/contacts');
     }
 
     /**
@@ -766,32 +766,48 @@ class ContactElement extends Element
     }
 
     /**
-     * Returns an image from unavatar.io, with an SVG behind it as a fallback.
+     * @inheritdoc
      */
-    public function getThumbHtml(int $size): ?string
+    public function getThumbUrl(int $size): ?string
     {
-        $value = strtolower(trim($this->email));
-        $image = Html::img('https://unavatar.io/' . $value . '?fallback=false', [
-            'width' => $size,
-            'height' => $size,
-            // Fixes a black border issue
-            'style' => 'border: 1px solid transparent',
-        ]);
+        $user = $this->getUser();
 
-        $initials = mb_strtoupper(mb_substr(str_replace('@', '', $value), 0, 2));
-        $svg = <<<XML
+        if ($user) {
+            return $user->getThumbUrl($size);
+        }
+
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function thumbSvg(): ?string
+    {
+        $initials = mb_strtoupper(mb_substr($this->email, 0, 2));
+
+        return <<<XML
             <svg version="1.1" baseProfile="full" width="100" height="100" viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
-                <circle cx="50" cy="50" r="50" fill="var(--gray-100)"/>
+                <circle cx="50" cy="50" r="50" fill="var(--gray-200)"/>
                 <text x="50" y="66" font-size="46" font-weight="500" font-family="sans-serif" text-anchor="middle" fill="var(--gray-700)">$initials</text>
             </svg>
         XML;
+    }
 
-        $options = [
-            'class' => 'elementthumb rounded',
-        ];
+    /**
+     * @inheritdoc
+     */
+    public function getThumbAlt(): ?string
+    {
+        return $this->email;
+    }
 
-        return Html::tag('div', $svg, $options)
-            . Html::tag('div', $image, $options);
+    /**
+     * @inheritdoc
+     */
+    public function getHasRoundedThumb(): bool
+    {
+        return true;
     }
 
     /**